Fixed DataManager and updated database scripts
Fixed problems in DataManager when creating tables, now sensor messages
should be stored properly. The issue was in the name "stored" used for
the datetime columns.

Updated the scripts used to initialize the databases, solving a problem
with missing table "arrowhead_service_interfaces", which had the name
changed from a previous version of Arrowhead framework.