Finished Ledger Integration
The connection to the Ledger is succesful. The connection to the ledger is split into two projects: One to upload a file to the Ledger (Ledger_PreConf) and the code to download the information is in the ERP_Consumer system.

Before running the demo for the first time it is required to upload the information to the ledger or the program will throw an exception and fail. The Ledger_Preconf is independent of the Arrowhead Framework and does not need any local cloud to be executed, just internet connection and the ledger configuration (crypto-config) files in the logged USER HOME directory